+++ src/useSQLite.ts
@@ -4,66 +4,105 @@ import { AvailableResult, notAvailable } from './util/models';
 import { isFeatureAvailable, featureNotAvailableError } 
                                     from './util/feature-check';
 import '@capacitor-community/sqlite';
+import { SQLiteDBConnection,
+         SQLiteConnection } from '@capacitor-community/sqlite';
 
-interface Set {
-    statement?: string;
-    values?: Array<any>;
+/**
+ * SQLite Hook Interface
+ */
+interface ISQLiteHook extends AvailableResult{
+    /**
+     * Echo a value
+     * @param value
+     * @returns Promise<{value: string}>
+     * @since 1.0.0 refactor
+     */
+    echo(value: string): Promise<{value: string}>;
+    /**
+     * Add an Upgrade Statement to Update Database Version
+     * @param dbName database name
+     * @param upgrade upgrade statement
+     * @returns Promise<Result>
+     * @since 1.0.0 refactor
+     */
+    addUpgradeStatement(dbName: string, upgrade: VersionUpgrade)
+                                                    :Promise<Result>;
+    /**
+     * Create a connection to a database
+     * @param database
+     * @param encrypted
+     * @param mode
+     * @param version
+     * @returns Promise<SQLiteDBConnection | Result | null>
+     * @since 1.0.0 refactor
+     */
+    createConnection(
+      database: string,
+      encrypted?: boolean,
+      mode?: string,
+      version?: number,
+    ): Promise<SQLiteDBConnection | Result | null>;
+    /**
+     * Retrieve an existing database connection
+     * @param database
+     * @returns Promise<SQLiteDBConnection | Result | null>
+     * @since 1.0.0 refactor
+     */
+    retrieveConnection(
+      database: string,
+    ): Promise<SQLiteDBConnection | Result | null>;
+    /**
+     * Retrieve all database connections
+     * @returns Promise<capSQLiteResult>
+     * @since 1.0.0 refactor
+     */
+    retrieveAllConnections(): Promise<any | null>;
+    /**
+     * Close a database connection
+     * @param database
+     * @returns Promise<Result>
+     * @since 1.0.0 refactor
+     */
+    closeConnection(database: string): Promise<Result>;
+    /**
+     * Close all database connections
+     * @returns Promise<Result>
+     * @since 1.0.0 refactor
+     */
+    closeAllConnections(): Promise<Result>;
+    /**
+     * Request Permissions
+     * @returns Promise<Result>
+     * @since 1.0.0 refactor
+     */
+    requestPermissions(): Promise<Result>;
+  
 }
-
+interface MySet {
+    statement?: string;
+    values?: any[];
+}
+  
 interface VersionUpgrade {
     fromVersion: number;
     toVersion: number;
     statement: string;
-    set?: Array<Set>; 
+    set?: MySet[]; 
 }
-
-interface SQLiteResult extends AvailableResult {
-    openDB: (dbName: string,encrypted?: boolean,mode?: string,
-            version?: number)
-                => Promise<{result?: boolean, message?: string}>;
-    createSyncTable: ()
-                => Promise<{changes?: {changes:number}}>;
-    close: (dbName: string)
-                => Promise<{result?: boolean, message?: string}>;
-    execute: (statements: string)
-                => Promise<{changes?: {changes: number},
-                            message?: string}>;
-    executeSet: (set: Array<Set>)
-                => Promise<{changes?: {changes: number, lastId: number},
-                            message?: string}>;
-    run: (statement: string,values?: Array<any>)
-                => Promise<{changes?: {changes: number, lastId: number},
-                            message?: string}>;
-    query: (statement: string,values?: Array<string>)
-                => Promise<{values?: Array<any>,message?: string}>
-    isDBExists: (dbName: string)
-                => Promise<{result?: boolean, message?: string}>;
-    deleteDB: (dbName: string)
-                => Promise<{result?: boolean, message?: string}>;
-    isJsonValid: (jsonstring: string)
-                => Promise<{result?: boolean, message?: string}>;
-    importFromJson: (jsonstring: string)
-                => Promise<{changes?: {changes: number},
-                            message?: string}>;
-    exportToJson: (mode: string)
-                => Promise<{export?: any,message?: string}>;
-    setSyncDate: (syncDate: string)
-                => Promise<{result?: boolean, message?: string}>;
-    addUpgradeStatement: (dbName: string, upgrade: VersionUpgrade)
-                => Promise<{result?: boolean, message?: string}>;
-    requestPermissions: () 
-                => Promise<{result?: boolean, message?: string}>;
+  
+interface Result {
+    result?: boolean;
+    message?: string
 }
-export const availableFeatures = {
-    useSQLite: isFeatureAvailable('CapacitorSQLite', 'useSQLite')
-}
+    
 /**
  * useSQLite Hook
  */
-export function useSQLite(): SQLiteResult {
+export function useSQLite(): ISQLiteHook {
     const { CapacitorSQLite } = Plugins;
     const platform = Capacitor.getPlatform();
-    const mSQLite: any = CapacitorSQLite;
+    const sqlitePlugin: any = CapacitorSQLite;
+    const mSQLite: SQLiteConnection = new SQLiteConnection(sqlitePlugin);
     let permissionsListener: any = null;
 
     const availableFeaturesN = {
@@ -71,7 +110,7 @@ export function useSQLite(): SQLiteResult {
     }
     const androidPermissions = async () => {
         try {
-            await mSQLite.requestPermissions();
+            await sqlitePlugin.requestPermissions();
             return { result: true };
         } catch (e) {
             console.log("Error requesting permissions " + e);
@@ -82,32 +121,25 @@ export function useSQLite(): SQLiteResult {
   
     if (!availableFeaturesN.useSQLite) {
         return {
-            openDB: featureNotAvailableError,
-            createSyncTable: featureNotAvailableError,
-            close: featureNotAvailableError,
-            execute: featureNotAvailableError,
-            executeSet: featureNotAvailableError,
-            run: featureNotAvailableError,
-            query: featureNotAvailableError,
-            isDBExists: featureNotAvailableError,
-            deleteDB: featureNotAvailableError,
-            isJsonValid: featureNotAvailableError,
-            importFromJson: featureNotAvailableError,
-            exportToJson: featureNotAvailableError,
-            setSyncDate: featureNotAvailableError,
+            echo: featureNotAvailableError,
+            createConnection: featureNotAvailableError,
+            closeConnection: featureNotAvailableError,
+            retrieveConnection: featureNotAvailableError,
+            retrieveAllConnections: featureNotAvailableError,
+            closeAllConnections: featureNotAvailableError,
             addUpgradeStatement: featureNotAvailableError,
             requestPermissions: featureNotAvailableError,
             ...notAvailable
         };
     }
     /**
-     * 
+     * Request Permissions
      */
     const requestPermissions = useCallback(async ():Promise<any> => {
         return new Promise(async (resolve) => {
             if(platform === "android") { 
 
-                permissionsListener = mSQLite.addListener(
+                permissionsListener = sqlitePlugin.addListener(
                         'androidPermissionsRequest',async (e: any) => {
                     if(e.permissionGranted === 0) {
                         permissionsListener.remove();
@@ -125,250 +157,98 @@ export function useSQLite(): SQLiteResult {
             }
         });
     }, []);
+    const echo = useCallback(async (value: string): Promise<any> => {
+        const r = await mSQLite.echo(value);
+        if(r) {
+            if( typeof r.value != 'undefined') {
+                return r;
+            }
+        } else {
+            return {value: null};
+        }
+    }, []);
     /**
-     * Open a Database
+     * Create a Connection to Database
      * @param dbName string
      * @param _encrypted boolean optional 
      * @param _mode string optional
      * @param version number optional
      */  
-    const openDB = useCallback(async (dbName: string,
-                                      encrypted?: boolean,
-                                      mode?: string,
-                                      version?: number): Promise<any> => {
-            console.log("%%% in openDB platform " + platform + "%%%");
+    const createConnection = useCallback(async (dbName: string,
+                                                encrypted?: boolean,
+                                                mode?: string,
+                                                version?: number)
+                        : Promise<SQLiteDBConnection| Result | null> => {
             if (typeof dbName === 'undefined') {
                 return { result: false,
                                 message: 'Must provide a database name'};
-            }      
+            } 
             const mDatabase: string = dbName;
             const mVersion: number = version ? version : 1;
             const mEncrypted: boolean = encrypted ? encrypted : false;
             const mMode: string = mode ? mode : "no-encryption";
-            const r = await mSQLite.open({database: mDatabase,
-                encrypted: mEncrypted,
-                mode: mMode, version: mVersion});
+            const r = await mSQLite.createConnection(
+                            mDatabase, mEncrypted, mMode, mVersion);
+
             if(r) {
-                if( typeof r.result != 'undefined') {
-                    return r;
-                }
-            }
-            return {result: false, message: "Error in openDB"};
-    }, []);
-    /**
-     * Create synchronisation table
-     */
-    const createSyncTable = useCallback(async () => {
-        const r = await mSQLite.createSyncTable();
-        console.log('result createSyncTable ',r);
-        if(r) {
-            if( typeof r.changes != 'undefined') {
                 return r;
             }
-        }
-        return {changes:0};
+            return null;
     }, []);
     /**
-     * Close the Database
+     * Close the Connection to the Database
      * @param dbName string
      */
-    const close = useCallback(async (dbName: string) => {
+    const closeConnection = useCallback(async (dbName: string) => {
         if(dbName.length > 0) {
-            const r = await mSQLite.close({database:dbName});
-            console.log('result close ',r);
+            const r = await mSQLite.closeConnection(dbName);
             if(r) {
                 if( typeof r.result != 'undefined') {
                     return r;
                 }
             } 
-            return {result: false, message: "Error in close"};  
+            return {result: false, message: "Error in closeConnection"};  
         }
         return {result: false, message: "Must provide a database name"};
     }, []);
     /**
-     * Execute a set of Raw Statements
-     * @param statements string 
-     */
-    const execute = useCallback(async (statements: string) => {
-        if(statements.length > 0) {
-            const r = await mSQLite.execute({statements:statements});
-            console.log('result execute ',r);
-            if(r) {
-                if( typeof r.changes != 'undefined') {
-                    return r;
-                }
-            } 
-            return {changes:{changes:0}, message: "Error in execute"};  
-        }
-        return {changes:{changes:0},message:"Statements is empty"};
-    }, []);
-    /**
-     * Execute a set of Raw Statements as Array<any>
-     * @param set Array<any> 
-     */
-    const executeSet = useCallback(async (set:Array<any>) => {
-        if(set.length > 0) {
-            const r = await mSQLite.executeSet({set:set});
-            console.log('result executeSet ',r);
-            if(r) {
-                if( typeof r.changes != 'undefined') {
-                    return r;
-                }
-            }           
-            return {changes:{changes: -1,lastId: -1},
-                                message: "Error in executeSet"};
-        }
-        return {changes:{changes:-1,lastId:-1},message:"Set is empty"};
-    }, []);
-    /**
-     * Execute a Single Raw Statement
-     * @param statement string
-     * @param values Array<any> optional
-     */
-    const run = useCallback(async (statement: string,
-                                  values?: Array<any>) => {
-        if(statement.length > 0) {
-            const vals: Array<any> = values ? values : [];
-            const r = await mSQLite.run({statement: statement,
-                                         values: vals});
-            console.log('result run ',r);
-            if(r) {
-                if( typeof r.changes != 'undefined') {
-                    return r;
-                }
-            } 
-            return {changes:{changes:0}, message: "Error in run"};  
-        }
-        return {changes:{changes:0,lastId:-1},
-                                message: "Statement is empty"};
-    }, []);
-    /**
-     * Query a Single Raw Statement
-     * @param statement string
-     * @param values Array<string> optional
-     */
-    const query = useCallback(async (statement: string,
-                                     values?:Array<string>) => {
-        if(statement.length > 0) {
-            const vals: Array<any> = values ? values : [];
-            const r = await mSQLite.query({statement: statement,
-                                           values: vals});
-            console.log('result query ',r);
-            if(r) {
-                if( typeof r.values != 'undefined') {
-                    return r;
-                }
-            } 
-            return {values:[], message: "Error in query"};  
-        }
-        return {values:[],message:"Statement is empty"};
-    }, []);
-    /**
-     * Check if the Database file exists
+     * Retrieve a Connection to the Database
      * @param dbName string
      */
-    const isDBExists = useCallback(async (dbName: string) => {
+    const retrieveConnection = useCallback(async (dbName: string) => {
         if(dbName.length > 0) {
-            const r = await mSQLite.isDBExists({database:dbName});
-            console.log('result isDBExists ',r);
+            const r = await mSQLite.retrieveConnection(dbName);
             if(r) {
-                if( typeof r.result != 'undefined') {
-                    return r;
-                }
+                return r;
             } 
-            return {result: false, message: "Error in isDBExists"};  
+            return null;  
         }
         return {result: false, message: "Must provide a database name"};
     }, []);
     /**
-     * Delete the Database file
+     * Retrieve all Connections to Databases
+     * 
+     */
+    const retrieveAllConnections = useCallback(async () => {
+            const r = await mSQLite.retrieveAllConnections();
+            var ret: any = {};
+            if(r) {
+               return r;
+            } 
+            return null;  
+    }, []);
+    /**
+     * Close All Connections to Databases
      * @param dbName string
      */
-    const deleteDB = useCallback(async (dbName: string) => {
-        if(dbName.length > 0) {
-            const r = await mSQLite.deleteDatabase({database:dbName});
-            console.log('result deleteDB ',r);
+    const closeAllConnections = useCallback(async () => {
+            const r = await mSQLite.closeAllConnections();
             if(r) {
                 if( typeof r.result != 'undefined') {
                     return r;
                 }
             } 
-            return {result: false, message: "Error in deleteDB"};  
-        }
-        return {result: false, message: "Must provide a database name"};
-    }, []);
-    /**
-     * Check the validity of a JSON Object
-     * @param jsonstring string 
-     */
-    const isJsonValid = useCallback(async (jsonstring: string) => {
-        if(jsonstring.length > 0) {
-            const r = await mSQLite.isJsonValid(
-                                        {jsonstring:jsonstring});
-            console.log('result isJsonValid ',r);
-            if(r) {
-                if( typeof r.result != 'undefined') {
-                    return r;
-                }
-            } 
-            return {result: false, message: "Error in isJsonValid"};  
-        }
-        return {result: false, message: "Must provide a Json string"};
-    }, []);
-    /**
-     * Import a database From a JSON Object
-     * @param jsonstring string 
-     */
-    const importFromJson = useCallback(async (jsonstring: string) => {
-        if(jsonstring.length > 0) {
-            const r = await mSQLite.importFromJson (
-                                        {jsonstring:jsonstring});
-            console.log('result importFromJson ',r);
-            if(r) {
-                if( typeof r.changes != 'undefined') {
-                    return r;
-                }
-            } 
-            return {changes:{changes:-1},
-                            message: "Error in importFromJson"};  
-        }
-        return {changes:{changes:-1},
-                            message: "Must provide a Json string"};
-    }, []);
-    /**
-     * Export the given database to a JSON Object
-     * @param mode string
-     */
-    const exportToJson = useCallback(async (mode: string) => {
-        if(mode.length > 0) {
-            const r = await mSQLite.exportToJson({jsonexportmode:mode});
-            console.log('result exportToJson ',r);
-            if(r) {
-                if( typeof r.export != 'undefined') {
-                    return r;
-                }
-            } 
-            return {export:{}, message: "Error in exportToJson"};  
-        }
-        return {export:{},message:"Must provide an export mode"};
-    }, []);
-    /**
-     * Set the synchronization date
-     * @param syncDate string 
-     */
-    const setSyncDate = useCallback(async (syncDate:string) => {
-        if(syncDate.length > 0) {
-            const r = await mSQLite.setSyncDate({syncdate:syncDate});
-            console.log('result setSyncDate ',r);
-            if(r) {
-                if( typeof r.result != 'undefined') {
-                    return r;
-                }
-            } 
-            return {result: false, message:"Error in setSyncDate"};
-        }
-        return {result: false,
-                    message:"Must provide a synchronization date"};
+            return {result: false, message: "Error in closeConnection"};  
     }, []);
     /**
      * Add the upgrade Statement for database version upgrading
@@ -387,23 +267,27 @@ export function useSQLite(): SQLiteResult {
                 msg += "fromVersion & toVersion & statement"
                 return {result: false,
                     message: msg};
-            }
+        }
 
         if(dbName.length > 0) {
-            const r = await mSQLite.addUpgradeStatement(
-                {database: dbName, upgrade: [upgrade]});
+            const r = await mSQLite
+                .addUpgradeStatement(dbName, upgrade.fromVersion,
+                                     upgrade.toVersion, upgrade.statement,
+                                     upgrade.set)
             if(r) {
                 if( typeof r.result != 'undefined') {
                     return r;
                 }
-            }  
+            } 
+            return {result: false,
+                    message:"addUpgradeStatement failed"};
         } else {
             return {result: false,
-                message:"Must provide a database name"};
+                    message:"Must provide a database name"};
         }
+
     }, []);
-    return { openDB, createSyncTable, close, execute, executeSet, run,
-        query, isDBExists, deleteDB, isJsonValid, importFromJson,
-        exportToJson, setSyncDate, addUpgradeStatement,
-        requestPermissions, isAvailable: true };
+    return { echo, createConnection, closeConnection, retrieveConnection,
+            retrieveAllConnections, closeAllConnections,
+            addUpgradeStatement, requestPermissions, isAvailable: true };
 }